+++ lofty/src/tag/items/mod.rs
@@ -1,6 +1,7 @@
 //! Various generic representations of tag items
 
 mod lang;
+pub mod popularimeter;
 mod timestamp;
 
 pub use lang::*;